CTF Compfest 17 Final/compfess/artifact/requirements.txt
django
gunicorn
whitenoise
psycopg2-binary
requests
urllib3
python-dotenv
Jinja2
djangorestframework
djangorestframework-simplejwt
django-extensions
bcrypt
argon2-cffi
